LIBITD-2592. Initial project setup
Initial setup of the "umd-handle-django" application following the
steps in https://umd-dit.atlassian.net/wiki/x/CAD8Ug

Used "umd_handle" for the project name, because using "umd-handle"
is not allowed (returned the following error):

```
CommandError: 'umd-handle' is not a valid project name. Please make sure the name is a valid identifier.
```

https://umd-dit.atlassian.net/browse/LIBITD-2592