LIBITD-2592. Initial REST API "api/v1/handles/<prefix>/<suffix> endpoint
Initial implementation of an "api/v1/handles/<prefix>/<suffix>"
REST endpoint that returns a JSON response with the URL of the handle,
or a 404 response if the specified handle is not found.

In the view, the “try-except” is necessary, because otherwise the
“get_object_or_404” method will return an HTML error page with an HTTP
Status 404 error code. In the Rails “umd-handle” application, an empty
JSON reponse (`{}`) was sent when there was a 404, and it seemed like a
good idea to continue that.

https://umd-dit.atlassian.net/browse/LIBITD-2592