LIBITD-2598. Added middleware to require login, except for specific URLs
Based on suggestions from Google's "AI Mode" and Stack Overflow, added
a "LoginRequiredMiddleware" class that would require a user to be
authenticated when going to any URL, except those starting with
specific prefixes.

The initial login URL and login callback URL are exempt, as well as
any URLs starting with "/api", as those are the REST API endpoints
that will use JWT for authnz.

https://umd-dit.atlassian.net/browse/LIBITD-2598