LIBITD-2598. Added "/health-check" endpoint that does not require auth
Added a minimal "/health-check" endpoint that does not require
authentication.

This is intended to be used with the Kubernetes readiness/liveness
probes.

Considered using an existing Python package such as
"django-health-check" or "django-watchman", but they appeared to
provide more functionality that was needed at the moment. For example,
both packages checked database connectivity by default, which is not
something is normally done in SSDR readiness/liveness probes/

https://umd-dit.atlassian.net/browse/LIBITD-2598