LIBITD-2600. Initial "JWTToken" model implementation
This is equivalent to the "JwtTokenLog" model in the Rails "umd-handle"
application.

Renamed to "JWTToken" because "JwtTokenLog" sounded more like a running
log of what JWT tokens has been used, vs. the list of actual JWT tokens
that were available for use.

Used "JWT" instead of "Jwt" in the class name, in keeping with Pep8
conventions when using "CapitalizedWords"

Ran `python src/manage.py makemigrations` to generate the migration
file.

https://umd-dit.atlassian.net/browse/LIBITD-2600