LIBITD-2604. Fixed JWT Authentication middleware issue
Fixed an issue with the “JWTAuthenticationMiddleware” from LIBITD-2600,
in which a new handle was being minted, even though no JWT was supplied
(or was invalid).

This was caused by a misunderstanding of how the Django middleware
works, specifically that when processing a request, the
`get_response` method must not be called until the JWT is
authenticated, i.e., from
https://docs.djangoproject.com/en/5.2/topics/http/middleware/:

```
    def __call__(self, request):
        # Code to be executed for each request before
        # the view (and later middleware) are called.

        response = self.get_response(request)

        # Code to be executed for each request/response after
        # the view is called.
```

Prior to the fix, the “get_response” was called at the start of the
method (and stored in a variable) which was later reused. Modified the
method to call “get_response” only when either the JWT token is not used
(for non-REST API calls) or after the JWT token has been verified.

This was not caught earlier because there were no REST API calls that
had a visible side effect (such as minting a new handle).

https://umd-dit.atlassian.net/browse/LIBITD-2604