LIBITD-2599. Environment banner for non-production environments
Selected the “django-admin-notice” package
(https://github.com/DoctorJohn/django-admin-notice) because it appears
to still be supported and the implementation was straightforward.

Added the “django-admin-notice” package to the “pyproject.toml” file.

The “django-admin-notice” package uses different settings than the
environment variables typically used in SSDR applications. In the
“settings.py” file, the standard SSDR environment variables are used to
provide the “django-admin-notice” settings, i.e.:

* ENVIRONMENT_BANNER - ADMIN_NOTICE_TEXT
* ENVIRONMENT_BANNER_FOREGROUND - ADMIN_NOTICE_TEXT_COLOR
* ENVIRONMENT_BANNER_BACKGROUND - ADMIN_NOTICE_BACKGROUND

The “django-admin-notice” package disables the banner when the
“ADMIN_NOTICE_TEXT” is empty or not set, therefore the
“ENVIRONMENT_BANNER_ENABLED” variable is not used.

In the “settings.py” file, the banner is configured for the local
development environment if the “DEBUG” setting is true (this is is
keeping with most other SSDR projects, where the local development
environment banner is displayed by default in “development” mode).
It is assumed that the Kubernetes environment will not run with the
“DEBUG” setting of “True”, and will configure the environment banner as
appropriate.

https://umd-dit.atlassian.net/browse/LIBITD-2599