LIBITD-2610. "exists" API endpoint check if handle exists
Added a GET "/api/v1/handles/exists" endpoint that returns
whether or not a handle with a given "repo" and "repo_id" exists.

Returns a 200 response and information about the handle when successful,
a 400 response (with an "errors" array) when parameters are missing or
invalid, and a 401 response when a JWT token is not provided or invalid.

https://umd-dit.atlassian.net/browse/LIBITD-2610