LIBITD-2611. "info" API endpoint for additional handle info
Added a GET "/api/v1/handles/info" endpoint that returns
whether or not a handle with a given "repo" and "repo_id" exists.

Returns a 200 response and information about the handle when successful,
a 400 response (with an "errors" array) when parameters are missing or
invalid, and a 401 response when a JWT token is not provided or invalid.

https://umd-dit.atlassian.net/browse/LIBITD-2611